requirements.txt
openai
discord.py
python-dotenv
colorizelib
pytube
requests
forex-python